ushadow/backend/requirements.txt
# ushadow Backend Requirements
# Python 3.11+

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.2.0

# HTTP Client for Chronicle/MCP/Agent Zero
httpx==0.27.2
aiohttp==3.11.7

# Database Clients
pymongo==4.9.0  # Motor 3.6.0 requires pymongo<4.10
motor==3.6.0  # Async MongoDB driver
redis==5.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
python-jose==3.3.0
bcrypt==4.2.1

# CORS & Middleware
fastapi-cors==0.0.6

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
PyYAML==6.0.2
docker==7.1.0  # Docker SDK for Python
sse-starlette==2.1.3  # Server-Sent Events for real-time Docker updates
qrcode[pil]==8.0  # QR code generation for Tailscale auth

# Logging & Monitoring
structlog==24.4.0

# YAML parsing for feature flags
ruamel.yaml==0.18.6

# Configuration Management
omegaconf==2.3.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2  # For testing
